requirements.txt
einops
requests
pytest
click
joblib
tqdm
ipykernel
notebook
ipywidgets
jupyter
flask
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
black
wandb
ftfy
tensorboard
pygame
plotly
sympy
accelerate
pre-commit
diffusers
gdown
jaxtyping
typeguard
streamlit
torchinfo
frozendict
streamlit-chat
pytorch-lightning
openai
transformer_lens
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
gym==0.23.1
pygame
torchtyping
protobuf==3.20.0
sentencepiece
transformers
huggingface_hub